requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.114.1
frozenlist==1.4.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1
greenlet==3.1.0
gspread==6.1.2
h11==0.14.0
idna==3.8
multidict==6.1.0
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic_core==2.23.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
yarl==1.11.1
